Added vim color scheme
Created a theme for vim. Is currently a semi-faithful one for one copy of the original to vim with minor changes made for new elements and elements which cannot be themed in the same way they are in Jetbrains IDEs. Added examples of the theme with syntax highlighting in both a full color GUI and a terminal with a 256 color space. Changed readme to reflect steps needed to use with vim. Theme currently works with full color terminals. As well as terminals limited to a 256, 16, 8, or 0 color space.